Fix: 코드 수정에 따른 TC 수정
